Widget settings for "total pages"
Add ability to align the totals overview and filter events

Also refactor the way widget settings work a bit; it was a bit kludgy
the way I did it yesterday.

+++ hit_test.go
@@ -167,8 +167,9 @@ func TestHitStatsList(t *testing.T) {
 				}
 			}
 
+			// TODO: add method to set this properly.
 			site.Settings.Widgets = goatcounter.Widgets{
-				{"on": true, "name": "pages", "s": map[string]interface{}{"limit_pages": 2, "limit_refs": 10}},
+				{"on": true, "name": "pages", "s": map[string]interface{}{"limit_pages": float64(2), "limit_refs": float64(10)}},
 			}
 
 			gctest.StoreHits(ctx, t, false, tt.in...)
